First pass iSH install

+++ user/local/bin/update-system.sh
@@ -1,9 +1,14 @@
 #!/usr/bin/env bash
 
-sudo apt update
-sudo apt full-upgrade
-sudo apt autoremove --purge --autoremove
-sudo apt clean
+if [[ "$HOSTNAME" == "Nathans-iPad" ]] && [[ -n "$(which apk)" ]]; then
+	apk update
+	apk upgrade
+elif [[ -n "$(which apt)" ]]; then
+	sudo apt update
+	sudo apt full-upgrade
+	sudo apt autoremove --purge --autoremove
+	sudo apt clean
+fi
 
 if [[ -n "$(which flatpak)" ]]; then
 	flatpak update
